Fix input variable parsing for TMVAGUI
TMVA::TMVAGlob::GetInputVariableNames wants to find the substring "__"
as part of the variable discovery process. However, it used
TString::First(char * character_set) which takes a char set as argument
and finds the first occurence of any character in that set.

This is now replaced by TString::Index which _can_ find a substring.